(Task): Fetching houses from Firestore completed
