+++ src/App.js
@@ -9,9 +9,9 @@ import Dropdown from "./Navbar/Dropdown";
 import Navbar from "./Navbar/Navbar";
 import GlobalStyle from "./globalStyles";
 import AuthForm from "./Footer/AuthForm";
-import AddNew from "./Components/AddNew";
+import AddNew from "./Components/Admin/AddNew";
 import NotFound from "./Components/NotFound";
-import { Admin } from "./Components/Admin";
+import { Admin } from "./Components/Admin/Admin";
 import { AuthProvider } from "./Footer/AuthContext";
 import PrivateRoute from "./PrivateRoute";
 import SingleHouse from "./Components/SingleHouse/SingleHouse";
